+++ bin/core/changes.md
@@ -2,9 +2,11 @@
 -----------
 
 * FIX: typing dot in non Haxe files
+* FIX: theme file switching on settings.json file save
 * CHANGED: do not change file using Navigation->Open File(completionItem.text = "", completionItem.displayText = "path/to/file")
 * CHANGED: scroll to bottom output textarea
 * CHANGED: Hxml completion: show completion on "-" only if it's at beginning of line
+* CHANGED: do not wrap file tree context menu items
 * ADDED: install missing haxelib dialog
 * ADDED: annotation ruler(with very basic style)
 * ADDED: basic outline panel